메뉴 리스트 정리 및 모달 연결 시도
메뉴 화면 백그라운드 이미지는 수정 예정
