+++ src/App.jsx
@@ -1,6 +1,6 @@
 //test code
 import Home from "./components/김경빈/Home";
-import MainHome from "./app/MainHome";
+import MainHome from "./app/mainHome";
 // import Topping from "./app/Modal";
 import { Routes, Route } from "react-router-dom";
 
@@ -9,7 +9,7 @@ function App() {
     <Routes>
       <Route path="/" element={<Home />} />
       <Route path="/test" element={<MainHome />} />
-      <Route path="/choice" element={<Topping />} />
+      {/* <Route path="/choice" element={<Topping />} /> */}
     </Routes>
   );
 }
