+++ src/components/menu/BurgerSetMenuData.jsx
@@ -1,0 +1,122 @@
+import { useState } from "react";
+
+function SetMenuData({ handleMenuClick }) {
+  const formatPrice = (price) => price.toLocaleString("ko-KR");
+
+  const [burgerSetMenuList] = useState([
+    {
+      id: 401,
+      name: "게살버거 세트",
+      price: 11000,
+      allergy: "갑각류, 마늘",
+      imgurl: "/image/buger/buger1.jpg",
+    },
+
+    {
+      id: 402,
+      name: "불고기 버거 세트",
+      price: 9500,
+      allergy: "마늘",
+      imgurl: "/image/buger/buger2.jpg",
+    },
+
+    {
+      id: 403,
+      name: "치킨버거 세트",
+      price: 11500,
+      allergy: "닭, 마늘",
+      imgurl: "/image/buger/buger3.jpg",
+    },
+
+    {
+      id: 404,
+      name: "쉬림프버거 세트",
+      price: 11500,
+      allergy: "새우, 마늘",
+      imgurl: "/image/buger/buger4.jpg",
+    },
+
+    {
+      id: 405,
+      name: "치즈 버거 세트",
+      price: 9500,
+      allergy: "유제품",
+      imgurl: "/image/buger/buger5.jpg",
+    },
+
+    {
+      id: 406,
+      name: "머쉬룸 버거 세트",
+      price: 11500,
+      allergy: "마늘",
+      imgurl: "/image/buger/buger6.jpg",
+    },
+
+    {
+      id: 407,
+      name: "핫치킨 버거 세트",
+      price: 10800,
+      allergy: "마늘, 닭",
+      imgurl: "/image/buger/buger7.jpg",
+    },
+
+    {
+      id: 408,
+      name: "라이스 버거 세트",
+      price: 9900,
+      allergy: "마늘",
+      imgurl: "/image/buger/buger8.jpg",
+    },
+
+    {
+      id: 409,
+      name: "더블치즈 버거 세트",
+      price: 9000,
+      allergy: "마늘, 유제품",
+      imgurl: "/image/buger/buger9.jpg",
+    },
+    {
+      id: 410,
+      name: "오징어 버거 세트",
+      price: 10800,
+      allergy: "마늘, 해산물, 오징어",
+      imgurl: "/image/buger/buger10.jpg",
+    },
+    {
+      id: 411,
+      name: "트러플치즈 버거 세트",
+      price: 12700,
+      allergy: "마늘",
+      imgurl: "/image/buger/buger11.jpg",
+    },
+
+    {
+      id: 412,
+      name: "돈까스 버거 세트",
+      price: 11800,
+      allergy: "마늘",
+      imgurl: "/image/buger/buger12.jpg",
+    },
+  ]);
+
+  return (
+    <div className="setmenu">
+      {/* 버거 세트 메뉴 리스트 */}
+      <ul>
+        {burgerSetMenuList.map((menu) => (
+          <li key={menu.id} onClick={() => handleMenuClick(menu)}>
+            <img
+              src={menu.imgurl}
+              alt={menu.name}
+              style={{ width: "100px", height: "100px" }}
+            />
+            <p>{menu.name}</p>
+            <p>가격: {formatPrice(menu.price)} 원</p>
+            <p>알레르기: {menu.allergy}</p>
+          </li>
+        ))}
+      </ul>
+    </div>
+  );
+}
+export default SetMenuData;
