+++ src/components/menu/Category.jsx
@@ -7,9 +7,9 @@ function Category({ setCurrentMenu }) {
       {/* 버튼 목록 생성 */}
       <button onClick={() => setCurrentMenu("burgerSetMenuList")}>세트</button>
       <button onClick={() => setCurrentMenu("burger")}>단품</button>
+      <button onClick={() => setCurrentMenu("side")}>사이드</button>
       <button onClick={() => setCurrentMenu("drink")}>음료</button>
       <button onClick={() => setCurrentMenu("coffee")}>커피</button>
-      <button onClick={() => setCurrentMenu("side")}>사이드</button>
     </div>
   );
 }
