+++ src/main.jsx
@@ -1,18 +1,20 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
+// import Home from "./components/김경빈/Home";
 // import "./styled/Modal.css";
-import Topping from "./app/Modal";
+// import Topping from "./app/Modal";
 // import App from "./components/topping/Counter.jsx";
 // import App from "./components/toppingslide";
 // import App from "/src/app/MainHome.jsx";
 // import App from "./App";
-// import App from "./App.jsx";
-import { BrowserRouter } from "react-router-dom";
+import App from "./App.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <BrowserRouter>
-      <Topping />
+      <App />
+      {/* <Topping /> */}
     </BrowserRouter>
   </StrictMode>
 );
